backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
PyGithub==1.59.1
python-gitlab==4.1.1
pydantic==2.5.0
aiofiles==23.2.1
httpx==0.25.2